requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
